+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### ✨ Improved
+
+* Add a couple checks to reject fits where either all the cameras have bad fit RMS (controlled by the `guider.max_fit_rms` parameters) or the delta scale value is too large (`guider.max_delta_scale_ppm`).
+
+
 ## 0.6.2 - April 28, 2023
 
 ### 🔧 Fixed